get logging handles working
